Conflicts:
	6LowPanManager/src/main/java/it/ismb/pertlab/pwal/wsn/driver/WSNAccelerometerSensor.java
	6LowPanManager/src/main/java/it/ismb/pertlab/pwal/wsn/driver/WSNTemperatureSensor.java
	bluetoothmanager/src/main/java/it/ismb/pertlab/pwal/manager/bluetooth/device/ThermometerDevice.java
	oxygenmeter-fake-driver/src/main/java/it/ismb/pertlab/pwal/manager/oxygenmeter/device/OxygenmeterFakeDevice.java
	pwal-api/src/main/java/it/ismb/pertlab/pwal/api/devices/interfaces/Device.java
	pwal-api/src/main/java/it/ismb/pertlab/pwal/api/devices/model/types/DeviceType.java
	pwal/pom.xml
	pwal/src/main/java/it/ismb/pertlab/pwal/App.java
	pwal/src/main/resources/applicationContext.xml
	thermometer-fake-driver/src/main/java/it/ismb/pertlab/pwal/manager/thermometer/device/ThermometerFakeDevice.java

+++ pwal-api/src/main/java/it/ismb/pertlab/pwal/api/devices/model/LightSensor.java
@@ -4,5 +4,5 @@ import it.ismb.pertlab.pwal.api.devices.interfaces.Device;
 
 public interface LightSensor extends Device{
 
-	Integer getLight();
+	Double getLight();
 }
